api de produto com lista e propriedade errada pro price... ta arredonddando

+++ grails-app/domain/muitofacilweb/Item.groovy
@@ -3,7 +3,7 @@ package muitofacilweb
 class Item {
 
 	String name
-	Long price
+	BigDecimal price
 	String obs
 	Long quantity
